+++ src/main/java/com/hencjo/summer/migration/Database.java
@@ -8,7 +8,7 @@ import java.sql.Statement;
 public final class Database {
 	public boolean containsTables(Connection connection) throws SQLException {
 		try (Statement s = connection.createStatement()) {
-			try (ResultSet r = s.executeQuery("SELECT count(*) as c FROM pg_tables WHERE schemaname='public';")) {
+			try (ResultSet r = s.executeQuery("SELECT count(*) as c FROM pg_tables WHERE schemaname=current_schema();")) {
 				r.next();
 				int numberOfTables = r.getInt(1);
 				return numberOfTables > 0;